+++ CHANGELOG.md
@@ -22,6 +22,8 @@ the detailed section referring to by linking pull requests or issues.
 
 #### Patch
 
+- Fixed gaps in renaming to "Data Offer"
+  ([#831](https://github.com/sovity/edc-ui/issues/831))
 - Replaced hints with info boxes in On Request data source
   ([#820](https://github.com/sovity/edc-ui/issues/820))
 - Fixed cropping of Contract Offer Ids on catalog browser page